FEA Make RAFT not be overwritten unles clean target is used

+++ python/setuputils.py
@@ -73,6 +73,8 @@ def use_raft_package(raft_path, cpp_build_path,
     """
     Function to use the python code in RAFT in package.raft
 
+    - If RAFT symling already exists, don't change anything. Use setup.py clean
+        if you want to change RAFT location.
     - Uses RAFT located in $RAFT_PATH if $RAFT_PATH exists.
     - Otherwise it will look for RAFT in the libcuml build folder,
         located either in the default location ../cpp/build or in
@@ -81,23 +83,28 @@ def use_raft_package(raft_path, cpp_build_path,
         - Branch/git tag cloned is located in git_info_file in this case.
 
     """
-    if not raft_path:
-        raft_path, raft_cloned = \
-            clone_repo_if_needed('raft', cpp_build_path,
-                                 git_info_file=git_info_file)
+    if not os.path.islink('cuml/raft'):
+        if not raft_path:
+            raft_path, raft_cloned = \
+                clone_repo_if_needed('raft', cpp_build_path,
+                                     git_info_file=git_info_file)
 
-        raft_path = '../' + raft_path
+            raft_path = '../' + raft_path
+
+        else:
+            print("-- Using RAFT_PATH variable, RAFT found at " +
+                  str(os.environ['RAFT_PATH']))
+            raft_path = os.environ['RAFT_PATH']
+
+        try:
+            os.symlink(raft_path + 'python/raft', 'cuml/raft')
+        except FileExistsError:
+            os.remove('cuml/raft')
+            os.symlink(raft_path + 'python/raft', 'cuml/raft')
 
     else:
-        print("-- Using RAFT_PATH variable, RAFT found at " +
-              str(os.environ['RAFT_PATH']))
-        raft_path = os.environ['RAFT_PATH']
-
-    try:
-        os.symlink(raft_path + 'python/raft', 'cuml/raft')
-    except FileExistsError:
-        os.remove('cuml/raft')
-        os.symlink(raft_path + 'python/raft', 'cuml/raft')
+        print("-- Using already existing RAFT folder, source located at")
+        print(os.path.realpath('cuml/raft'))
 
 
 def clone_repo_if_needed(name, cpp_build_path,
@@ -162,11 +169,11 @@ def get_submodule_dependency(repo,
 
     else:
 
-        warnings.warn("-- Third party repositories have not been found so they"
-                      "will be cloned. To avoid this set the environment "
-                      "variable CUML_BUILD_PATH, containing the relative "
-                      "path of the root of the repository to the folder "
-                      "where libcuml++ was built.")
+        print("-- Third party repositories have not been found so they " +
+              "will be cloned. To avoid this set the environment " +
+              "variable CUML_BUILD_PATH, containing the relative " +
+              "path of the root of the repository to the folder " +
+              "where libcuml++ was built.")
 
         for repo in repos:
             clone_repo(repo, repo_info[repo][0], repo_info[repo][1])
